# openapi.yaml
openapi: 3.0.3
info:
  title: Open Mic API
  version: 1.0.0
  contact:
    email: info@openmics.org
  description:
    Canonical API contract. All paths are relative to the /api base URL.
    The public media deep-link page (`/media/:mediaId`) is intentionally not part of
    this document — like the `/@:handle` vanity resolver it is served outside `/api`
    as server-rendered SPA HTML with Open Graph tags, not as a JSON operation.
servers:
  - url: https://api.openmics.org/api
    description: Production
...
paths:
...
  /auth/profile:
    get:
      tags:
        - Auth
      operationId: getAuthProfile
      description: Returns the authenticated caller's authentication profile.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me:
    get:
      tags:
        - Accounts
      operationId: getMe
      description: Returns the account associated with the authenticated caller.
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
  /me/permissions:
    get:
      tags:
        - Accounts
      operationId: getMyPermissions
      description: Returns the authenticated caller's permissions, optionally scoped to a profile.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProfileQuery"
      responses:
        "200":
          $ref: "#/components/responses/PermissionResponse"
  /accounts/{id}:
    patch:
      tags:
        - Accounts
      operationId: updateAccount
      description: Partially updates the account identified by `id`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        $ref: "#/components/requestBodies/AccountUpdate"
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
  /accounts/{id}/profiles:
    get:
      tags:
        - Accounts
      operationId: listAccountProfiles
      description: Lists the profiles associated with the account identified by `id`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/ProfileListResponse"
  /accounts/{id}/current-profile:
    put:
      tags:
        - Accounts
      operationId: setCurrentProfile
      description: Sets the selected current profile for the account identified by `id`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        $ref: "#/components/requestBodies/CurrentProfile"
      responses:
        "200":
          $ref: "#/components/responses/ProfileResponse"
...
  /cities/search:
    get:
      tags:
        - Cities
      operationId: searchCities
      description: Searches the bounded local city catalogue using Unicode and ASCII city/country spellings. Results rank exact names, then prefixes, then other matches, with population descending within each tier.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: country
          in: query
          required: false
          description: Optional ISO 3166-1 alpha-2 country context.
          schema:
            type: string
            pattern: "^[A-Za-z]{2}$"
      responses:
        "200":
          description: At most ten matching catalogue cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitySearchResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /cities/{id}:
    get:
      tags:
        - Cities
      operationId: getCity
      description: Resolves a saved city ID to its canonical catalogue names, country, region and centre coordinates.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Resolved city
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /cities/search-external:
    post:
      tags:
        - Cities
      operationId: searchExternalCities
      description: Explicit city-only LocationIQ lookup. Successful responses, including empty results, are cached for 24 hours and confirmed identities are persisted in the local catalogue. Provider use requires a configured finite GEOCODING_DAILY_LIMIT; cache hits do not consume the shared daily budget.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [q]
              additionalProperties: false
              properties:
                q:
                  type: string
                  minLength: 2
                  maxLength: 100
      responses:
        "200":
          description: Provider city results, or a successful cache hit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalCitySearchResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/GeocodingRateLimited"
        "503":
          $ref: "#/components/responses/GeocodingUnavailable"
  /discovery/suggestions:
    get:
      tags:
        - Discovery
      operationId: getDiscoverySuggestions
      description: Returns the smallest outward-rounded radius containing twenty or more additional active public series outside the current radius (or the positive remaining count at 200 km), plus up to six other catalogue cities with active public series ordered by centre distance. Hidden series are excluded for every caller.
      parameters:
        - name: near
          in: query
          required: true
          description: Origin coordinates formatted as lat,lng.
          schema:
            type: string
            pattern: "^-?[0-9]+(?:\\.[0-9]+)?,-?[0-9]+(?:\\.[0-9]+)?$"
        - name: radius_km
          in: query
          required: false
          schema:
            type: number
            minimum: 0
            maximum: 200
            default: 50
        - name: city_id
          in: query
          required: false
          description: Current resolved city to exclude from alternative city suggestions.
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: Expansion and city suggestion metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverySuggestions"
        "400":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    editSession:
      type: apiKey
      in: cookie
      name: openmic_edit_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ProfileQuery:
      name: profile
      in: query
      schema:
        type: string
        format: uuid
...
  requestBodies:
    AccountUpdate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AccountUpdateRequest"
    CurrentProfile:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CurrentProfileRequest"
...
  responses:
    AccountResponse:
      description: Account
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Account"
    ProfileResponse:
      description: Profile
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Profile"
...
    PermissionResponse:
      description: Permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PermissionSet"
...
    ProfileListResponse:
      description: Profiles
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProfilePage"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    GeocodingRateLimited:
      description:
        The geocoding provider (or the server-side proxy's shared rate
        limit) rejected the request; back off and retry later, or continue without
        geocoding assist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    GeocodingUnavailable:
      description: Geocoding is not configured on this server (no provider API key set).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Id:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    ProfileVisibility:
      type: string
      enum:
        - public
        - unlisted
        - private
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
              description:
                Machine-readable extras specific to `code` — validation field
                errors for VALIDATION_ERROR, or a `scope` discriminator for
                PLAN_LIMIT_EXCEEDED / MEDIA_QUOTA_EXCEEDED.
...
    City:
      type: object
      required:
        - id
        - city
        - city_ascii
        - country
        - country_ascii
        - iso2
        - iso3
        - admin_name
        - lat
        - lng
        - population
      properties:
        id:
          $ref: "#/components/schemas/Id"
        city:
          type: string
        city_ascii:
          type: string
        country:
          type: string
        country_ascii:
          type: string
        iso2:
          type: string
          pattern: "^[A-Z]{2}$"
        iso3:
          type: string
          pattern: "^[A-Z]{3}$"
          nullable: true
        admin_name:
          type: string
          nullable: true
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
        population:
          type: integer
          minimum: 0
          nullable: true
    CitySearchResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/City"
    ExternalCitySearchResponse:
      type: object
      required: [items, cached]
      properties:
        items:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/City"
        cached:
          type: boolean
    DiscoverySuggestions:
      type: object
      required: [expansion, cities]
      properties:
        expansion:
          type: object
          nullable: true
          required: [radius_km, additional_count]
          properties:
            radius_km:
              type: number
              minimum: 0
              maximum: 200
            additional_count:
              type: integer
              minimum: 1
        cities:
          type: array
          maxItems: 6
          items:
            allOf:
              - $ref: "#/components/schemas/City"
              - type: object
                required: [open_mic_count, distance_km]
                properties:
                  open_mic_count:
                    type: integer
                    minimum: 1
                  distance_km:
                    type: number
                    minimum: 0
                  approximate:
                    type: boolean
    Account:
      type: object
      required:
        - id
        - email
        - display_name
        - city_id
        - city_location
        - plan
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/Id"
        cognito_id:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        city:
          type: string
          nullable: true
        city_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
        city_location:
          allOf:
            - $ref: "#/components/schemas/City"
          nullable: true
        preferred_language:
          type: string
          nullable: true
        current_profile_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
        is_platform_admin:
          type: boolean
        plan:
          type: string
          enum:
            - free
            - pro
        referred_by_profile_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
        referred_at:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          nullable: true
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    Profile:
      type: object
      required:
        - id
        - created_by_account_id
        - profile_name
        - visibility
        - current_handle
      properties:
        id:
          $ref: "#/components/schemas/Id"
        created_by_account_id:
          $ref: "#/components/schemas/Id"
        current_handle:
          type: string
          nullable: true
          description: Always null for organizer profiles; organizers are represented publicly by their open-mic series handles.
        profile_name:
          type: string
        profile_kind:
          type: string
          enum:
            - organizer
            - performer
        bio:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
          description: Private contact phone for the profile owner and permitted organizers; never shown in public profile reads.
        profile_image_url:
          type: string
          format: uri
          nullable: true
        theme_name:
          type: string
          nullable: true
        color_mode:
          type: string
          enum:
            - light
            - dark
          nullable: true
        visibility:
          $ref: "#/components/schemas/ProfileVisibility"
        show_gig_media:
          type: boolean
          description:
            Defaults to true. Controls only the derived media gallery on this
            profile's public page; event and series galleries are unaffected.
        links:
          type: array
          items:
            $ref: "#/components/schemas/ProfileLink"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
        deleted_at:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          nullable: true
        recovery_deadline:
          allOf:
            - $ref: "#/components/schemas/Timestamp"
          nullable: true
    ProfileLink:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        profile_id:
          $ref: "#/components/schemas/Id"
        type:
          type: string
          enum:
            - website
            - instagram
            - youtube
            - tiktok
            - twitter
            - facebook
            - spotify
            - bandcamp
            - email
            - phone
            - other
        url:
          type: string
        label:
          type: string
          nullable: true
        sequence:
          type: integer
...
    PermissionSet:
      type: object
      properties:
        permissions:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
    ProfilePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    AccountUpdateRequest:
      type: object
      properties:
        display_name:
          type: string
        city:
          type: string
          nullable: true
        city_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          nullable: true
        preferred_language:
          type: string
    CurrentProfileRequest:
      type: object
      required:
        - profile_id
      properties:
        profile_id:
          $ref: "#/components/schemas/Id"
...